% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playVoice.R
\name{playVoice}
\alias{playVoice}
\title{playVoice
This function takes the Wave object and plays only the speech parts, segment by segment, using tuneR::play() function.}
\usage{
playVoice(x, seg = NA)
}
\arguments{
\item{x}{The input Wave object}

\item{seg}{Segmentation obtained by voiceSegment(..,unit="time")}
}
\description{
playVoice
This function takes the Wave object and plays only the speech parts, segment by segment, using tuneR::play() function.
}
\examples{
\dontrun{
library(tuneR)
x <- readWave("speech.wav")
playVoice(x)
}
}
