% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voiceActivity.R
\name{voiceActivity}
\alias{voiceActivity}
\title{voiceActivity
This function takes the Wave object and returns an array of the voice activity.}
\usage{
voiceActivity(x, minlen = 20, nclust = 3, frameshift = 0.01)
}
\arguments{
\item{x}{The input Wave object}

\item{minlen}{Minimum duration of silence (number of frames)}

\item{nclust}{Number of clusters, one of which should be silent}

\item{frameshift}{The frame shift (sec), 10ms by default}
}
\value{
A vector of logical with length of number of frames. If the frame is the voice frame, the content of the vector is TRUE, otherwise FALSE.
}
\description{
voiceActivity
This function takes the Wave object and returns an array of the voice activity.
}
\examples{
\dontrun{
library(tuneR)
x <- readWave("speech.wav")
act <- voiceActivity(x)
}
}

